...G...... [ZBX-20860] changed rdev variable to be cast to uint64 in dev_linux.go

+++ src/go/plugins/vfs/dev/dev_linux.go
@@ -69,7 +69,7 @@ func (p *Plugin) getDiscovery() (out string, err error) {
 			if stat.Mode()&os.ModeType == os.ModeDevice {
 				dev := &devRecord{Name: entry.Name()}
 				if sysfs {
-					rdev := stat.Sys().(*syscall.Stat_t).Rdev
+					rdev := uint64(stat.Sys().(*syscall.Stat_t).Rdev)
 					dirname := fmt.Sprintf("%s%d:%d/", sysBlkdevLocation, unix.Major(rdev), unix.Minor(rdev))
 
 					if lstat, tmperr := os.Lstat(devname); tmperr == nil {
@@ -145,7 +145,7 @@ func (p *Plugin) getDeviceName(name string) (devName string, err error) {
 		if minor, err = strconv.ParseUint(fields[1], 10, 32); err != nil {
 			return
 		}
-		rdev := stat.Sys().(*syscall.Stat_t).Rdev
+		rdev := uint64(stat.Sys().(*syscall.Stat_t).Rdev)
 		if uint64(unix.Major(rdev)) == major && uint64(unix.Minor(rdev)) == minor {
 			return fields[2], nil
 		}
@@ -190,7 +190,7 @@ func (p *Plugin) scanDeviceStats(name string, buf *bytes.Buffer) (devstats *devS
 		}
 		var stat os.FileInfo
 		if stat, err = os.Stat(name); err == nil {
-			rdev = stat.Sys().(*syscall.Stat_t).Rdev
+			rdev = uint64(stat.Sys().(*syscall.Stat_t).Rdev)
 		}
 	}
